feat: game connection stage & second player connection establishment
