+++ src/libs/config.js
@@ -1,19 +1,21 @@
 export const GAME_STAGES = {
   menu: 0,
-  planning: 1,
-  planningComplete: 2,
-  ready: 3,
-  ongoing: 4,
-  gameover: 5,
+  connection: 1,
+  planning: 2,
+  planningComplete: 3,
+  ready: 4,
+  ongoing: 5,
+  gameover: 6,
 };
 
 export const GAME_STAGE_MAP = {
   0: 'Menu',
-  1: 'Planning',
-  2: 'Planning Complete',
-  3: 'Ready',
-  4: 'Game',
-  5: 'gameover',
+  1: 'Connection',
+  2: 'Planning',
+  3: 'Planning Complete',
+  4: 'Ready',
+  5: 'Game',
+  6: 'gameover',
 };
 
 export const SHIP_DETAILS = [
@@ -96,4 +98,8 @@ export const INITIAL_BATTLEFIELD_SETUP = {
 export const BATTLEFIELD_SIDES = {
   player: 'player',
   enemy: 'enemy',
+};
+
+export const ENV_VARS = {
+  socketPath: process.env.SOCKET_MULTIPLAYER_PATH || '/api/multiplayer_connection',
 };
